Navbar fixes.
Most nabber items should now be hidden if you do not have the required
role.
This does not apply to the manager bit cos we still need to work out
who classes as a manager.

Things that still need to be done.

Authentication check when browsing to specific URL’s that you are not
authorized for its proper error handling.